Changed the type of hashes to size_t, mostly since many operations that are undefined on signed integer types are defined for unsigned integer types like size_t. Made integer truncation work as expected in Object_cmp, if the resulting integer difference is too large to be held in an int we truncate the value to the closest number representable by an int. Added private/memory.h include to private/memory.c in we change the signature and don't update the implementation, updated Makefile to reflect this.

+++ base/Metaclass/Metaclass.c
@@ -111,7 +111,7 @@ size_t coal_base_Object_getSize (val self) {
   return class->size;
 }
 
-int coal_base_Object_hashCode (val self) {
+size_t coal_base_Object_hashCode (val self) {
   const class(Metaclass) * class;
 
   class = coal_base_Object_getClass(self);
